Rework provider to use chain_err()
